+++ karma.conf.js
@@ -32,7 +32,13 @@ module.exports = function (config) {
       },
     },
     reporters: ['progress', 'kjhtml'],
-    browsers: ['Chrome'],
+    customLaunchers: {
+      ChromeHeadlessCI: {
+        base: 'ChromeHeadless',
+        flags: ['--no-sandbox', '--disable-gpu', '--disable-dev-shm-usage'],
+      },
+    },
+    browsers: ['ChromeHeadlessCI'],
     restartOnFileChange: true,
   });
 };