+++ projects/demo/src/app/app.routes.ts
@@ -1,14 +1,9 @@
 import type { Routes } from '@angular/router';
-import { UaePassCallbackComponent } from 'sensei-uaepass';
 
 export const routes: Routes = [
   {
     path: '',
-    loadComponent: () => import('./home/home.component').then((m) => m.HomeComponent),
-  },
-  {
-    path: 'uae-pass/callback',
-    component: UaePassCallbackComponent,
+    loadComponent: () => import('./home/home.component').then((module) => module.HomeComponent),
   },
   {
     path: '**',