+++ projects/uae-pass/src/lib/uae-pass.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'lib-uae-pass',
-  imports: [],
-  template: ` <p>uae-pass works!</p> `,
-  styles: ``,
-})
-export class UaePassComponent {}