+++ projects/uae-pass/src/public-api.ts
@@ -4,14 +4,12 @@
 
 // Configuration and types
 export * from './lib/uae-pass.config';
-export * from './lib/uae-pass.constants';
 export * from './lib/uae-pass.types';
 export * from './lib/uae-pass.enums';
 export * from './lib/uae-pass.error';
 
 // Core service
-export * from './lib/uae-pass.oauth.service';
+export * from './lib/uae-pass-auth.service';
 
 // UI components
 export * from './lib/uae-pass-login-button.component';
-export * from './lib/uae-pass-callback.component';